add link subcommand
